navigation from goal index to goal form

+++ shared/src/commonMain/kotlin/com/nabeelkm/workout/entity/Goal.kt
@@ -2,6 +2,7 @@ package com.nabeelkm.workout.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import org.koin.core.logger.Logger
 
 enum class GoalStatus(
     val value: Int,
@@ -16,9 +17,9 @@ data class Goal(
     @PrimaryKey(autoGenerate = true) val id: Int,
     val name: String,
     val status: Int,
-    val createdAt: Int,
-    val completedAt: Int?,
-    val startAt: Int?
+    val createdAt: Long,
+    val completedAt: Long?,
+    val startAt: Long?
 ) {
     fun getStatus(): GoalStatus {
         return GoalStatus.entries[status]
